// src/models/reservation-create-req.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 预定创建请求
 * @export
 * @interface ReservationCreateReq
 */
export interface ReservationCreateReq {
    /**
     * 门店ID
     * @type {number}
     * @memberof ReservationCreateReq
     */
    storeId: number;
    /**
     * 房间ID
     * @type {number}
     * @memberof ReservationCreateReq
     */
    roomId: number;
    /**
     * 会员ID
     * @type {number}
     * @memberof ReservationCreateReq
     */
    memberId: number;
    /**
     * 申请人(业务员)ID
     * @type {number}
     * @memberof ReservationCreateReq
     */
    staffId: number;
    /**
     * 预定日期 yyyy-MM-dd
     * @type {string}
     * @memberof ReservationCreateReq
     */
    reserveDate: string;
    /**
     * 预计人数
     * @type {number}
     * @memberof ReservationCreateReq
     */
    guestCount?: number;
    /**
     * 备注
     * @type {string}
     * @memberof ReservationCreateReq
     */
    remark?: string;
}