// src/models/room-schedule-booking-resp.ts
// tslint:disable
/**
 * geo
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 单日房态
 * @export
 * @interface RoomScheduleBookingResp
 */
export interface RoomScheduleBookingResp {
    /**
     * 房态: AVAILABLE/PENDING/BOOKED/FINISHED
     * @type {string}
     * @memberof RoomScheduleBookingResp
     */
    state?: string;
    /**
     * 预订ID
     * @type {number}
     * @memberof RoomScheduleBookingResp
     */
    reservationId?: number;
    /**
     * 预订单号
     * @type {string}
     * @memberof RoomScheduleBookingResp
     */
    reserveNo?: string;
    /**
     * 预订状态(PENDING/APPROVED)
     * @type {string}
     * @memberof RoomScheduleBookingResp
     */
    status?: RoomScheduleBookingRespStatusEnum;
    /**
     * 会员ID
     * @type {number}
     * @memberof RoomScheduleBookingResp
     */
    memberId?: number;
    /**
     * 业务员ID
     * @type {number}
     * @memberof RoomScheduleBookingResp
     */
    staffId?: number;
}

/**
    * @export
    * @enum {string}
    */
export enum RoomScheduleBookingRespStatusEnum {
    PENDING = 'PENDING',
    APPROVED = 'APPROVED',
    REJECTED = 'REJECTED',
    CANCELLED = 'CANCELLED'
}